# _docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PlayMyDate API
  description: |
    PlayMyDate v2 API documentation for B2C and B2B features.

    ## Authentication
    - **B2C**: Session-based authentication via Lucia
    - **B2B**: Session-based authentication with merchant credentials

    ## Rate Limiting
    - Standard: 100 requests/minute
    - Premium: 500 requests/minute
  version: 2.0.0
  contact:
    name: PlayMyDate Support
servers:
  - url: /api
    description: API server
...
paths:
...
  /playcoin/exchange:
    post:
      tags: [PlayCoin]
      summary: Exchange coins with partner
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [partnerId, amount]
              properties:
                partnerId:
                  type: string
                amount:
                  type: integer
                  minimum: 1
      responses:
        '200':
          description: Exchange result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExchangeResponse'

  # Dates/Paradise Mode
...
  /subscription/upgrade:
    post:
      tags: [Subscription]
      summary: Upgrade subscription tier
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tier]
              properties:
                tier:
                  type: string
                  enum: [plus, premium]
      responses:
        '200':
          description: Upgrade result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
...
  /b2b/analytics/overview:
    get:
      tags: [B2B Analytics]
      summary: Get analytics overview
      security:
        - b2bSessionAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d, all]
            default: 30d
      responses:
        '200':
          description: Analytics overview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsOverviewResponse'
  /b2b/analytics/conversions:
    get:
      tags: [B2B Analytics]
      summary: Get conversion funnel
      security:
        - b2bSessionAuth: []
      parameters:
        - name: campaignId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Conversion data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversionFunnelResponse'
  /b2b/analytics/audience:
    get:
      tags: [B2B Analytics]
      summary: Get audience insights
      security:
        - b2bSessionAuth: []
      responses:
        '200':
          description: Audience data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AudienceInsightsResponse'
  /b2b/analytics/reports:
    post:
      tags: [B2B Analytics]
      summary: Generate report
      security:
        - b2bSessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateReportRequest'
      responses:
        '201':
          description: Report generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
...
  /b2b/billing/invoices:
    get:
      tags: [B2B Billing]
      summary: List invoices
      security:
        - b2bSessionAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, paid, overdue]
      responses:
        '200':
          description: Invoice list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoicesResponse'
  /b2b/billing/usage:
    get:
      tags: [B2B Billing]
      summary: Get usage data
      security:
        - b2bSessionAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d]
            default: 30d
      responses:
        '200':
          description: Usage data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageResponse'
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: auth_session
    b2bSessionAuth:
      type: apiKey
      in: cookie
      name: b2b_session
  schemas:
...
    ExchangeResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            transactionId:
              type: string
            partnerId:
              type: string
            partnerName:
              type: string
            coinsSpent:
              type: integer
            rewardValue:
              type: integer
            code:
              type: string
...
    SubscriptionResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            userId:
              type: string
              format: uuid
            tier:
              type: string
              enum: [free, plus, premium]
            status:
              type: string
            currentPeriodEnd:
              type: string
              format: date-time
...
    AnalyticsOverviewResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            totalCampaigns:
              type: integer
            activeCampaigns:
              type: integer
            totalImpressions:
              type: integer
            totalClicks:
              type: integer
            totalConversions:
              type: integer
            totalSpent:
              type: number
            avgCTR:
              type: number
            avgConversionRate:
              type: number
    ConversionFunnelResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            impressions:
              type: integer
            clicks:
              type: integer
            engagements:
              type: integer
            conversions:
              type: integer
            dropoffRates:
              type: object
    AudienceInsightsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            type: object
            properties:
              segment:
                type: string
              count:
                type: integer
              percentage:
                type: number
              avgEngagement:
                type: number
    GenerateReportRequest:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [overview, audience, campaigns]
        period:
          type: string
          enum: [7d, 30d, 90d, all]
        campaignIds:
          type: array
          items:
            type: string
            format: uuid
    ReportResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
            reportType:
              type: string
            status:
              type: string
            data:
              type: object
            generatedAt:
              type: string
              format: date-time
    InvoicesResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Invoice'
    Invoice:
      type: object
      properties:
        id:
          type: string
          format: uuid
        invoiceNumber:
          type: string
        amount:
          type: number
        status:
          type: string
        dueDate:
          type: string
          format: date
        paidAt:
          type: string
          format: date-time
    UsageResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            totalImpressions:
              type: integer
            totalClicks:
              type: integer
            totalConversions:
              type: integer
            totalSpend:
              type: number
            dailyUsage:
              type: array
              items:
                type: object