# src/app/api/b2b/analytics/conversions/route.ts
import { NextRequest, NextResponse } from 'next/server';
import { withB2BAuth } from '@/lib/api-auth-b2b';
import { getConversionFunnel } from '@/lib/services/analytics';

export async function GET(request: NextRequest) {
  return withB2BAuth(request, async () => {
    const { searchParams } = new URL(request.url);
    const campaignId = searchParams.get('campaignId');

    if (!campaignId) {
      return NextResponse.json({ error: 'campaignId is required' }, { status: 400 });
    }

    const result = await getConversionFunnel(campaignId);

    if (!result.success) {
      return NextResponse.json({ error: result.error }, { status: 500 });
    }

    return NextResponse.json(result.data);
  });
}